separate project/server + class features
