+++ activetigger/functions.py
@@ -3,6 +3,7 @@ from pandas import DataFrame, Series
 import fasttext
 import spacy
 from sentence_transformers import SentenceTransformer
+
 from sklearn.preprocessing import StandardScaler
 from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
 from sklearn.linear_model import LogisticRegression
@@ -10,12 +11,12 @@ from sklearn.naive_bayes import MultinomialNB
 from sklearn.naive_bayes import BernoulliNB
 from sklearn.ensemble import RandomForestClassifier
 from sklearn.neighbors import KNeighborsClassifier
-
 from sklearn.metrics import precision_score
 
 
-
-# tokenisation ?
+# nexts steps : 
+# - connect simplemodel to server
+# - deal multiple schemes
 
 
 class SimpleModel():
@@ -31,24 +32,45 @@ class SimpleModel():
                  standardize:bool=False,
                  **kwargs
                  ):
+        """
+        Initialize simpe model for data
+        model (str): type of models
+        data (DataFrame): dataset
+        label (str): column of the tags
+        predictor (list): columns of predictors
+        standardize (bool): predictor standardisation
+        **kwargs: parameters for models
+        """
         
         self.available_models = ["simplebayes",
                                  "liblinear",
                                  "knn",
                                  "randomforest",
                                  "lasso"]
-
-        # Data curation
         
-        # Drop NaN
-        df = data[[label]+predictors].dropna()
+        self.col_label = label
+        self.col_predictors = predictors
 
-        self.Y = df[label]
-        self.X = df[predictors]
-        self.labels = self.Y.unique()
+        # Build the data set & missing predictors
+        # For the moment remove missing predictors
+        f_na = data[predictors].isna().sum(axis=1)>0
+        
+        if f_na.sum()>0:
+            print(f"There is {f_na.sum()} predictor rows with missing values")
 
         if standardize:
-            self.standardize()
+            df_pred = self.standardize(data[~f_na][predictors])
+        else:
+            df_pred = data[~f_na][predictors]
+
+        self.df = pd.concat([data[~f_na][label],df_pred],axis=1)
+        
+        # data for training
+        f_label = self.df[self.col_label].notnull()
+        self.Y = self.df[f_label][label]
+        self.X = self.df[f_label][predictors]
+
+        self.labels = self.Y.unique()
 
         # Select model
         if model == "knn":
@@ -66,10 +88,15 @@ class SimpleModel():
                                             solver="liblinear",
                                             C = C)
         if model == "naivebayes":
+            if not "distribution" in kwargs:
+                raise TypeError("Missing distribution parameter for naivebayes")
+            
         # only dfm as predictor
             alpha = 1
             if "smooth" in kwargs:
                 alpha = kwargs["smooth"]
+            fit_prior = True
+            class_prior = None
             if "prior" in kwargs:
                 if kwargs["prior"] == "uniform":
                     fit_prior = True
@@ -80,6 +107,7 @@ class SimpleModel():
                 if kwargs["prior"] == "termfreq":
                     fit_prior = False
                     class_prior = None #TODO
+ 
             if kwargs["distribution"] == "multinomial":
                 self.model = MultinomialNB(alpha=alpha,
                                            fit_prior=fit_prior,
@@ -114,17 +142,20 @@ class SimpleModel():
             self.model = RandomForestClassifier(n_estimators=100, 
                                                 random_state=42,
                                                 max_features=max_features)
-
+        
         # Fit model
         self.model.fit(self.X, self.Y)
+        self.proba = pd.DataFrame(self.predict_proba(self.df[self.col_predictors]),
+                                 columns = self.model.classes_)
         self.precision = self.compute_precision()
 
-    def standardize(self):
+    def standardize(self,df):
         """
         Apply standardization
         """
         scaler = StandardScaler()
-        self.X = scaler.fit_transform(self.X)
+        df_stand = scaler.fit_transform(df)
+        return pd.DataFrame(df_stand,columns=df.columns,index=df.index)
 
     def compute_precision(self):
         """
@@ -132,8 +163,22 @@ class SimpleModel():
         """
         y_pred = self.model.predict(self.X)
         precision = precision_score(list(self.Y), list(y_pred),pos_label=self.labels[0])
-        return precision
-
+        return 
+    
+    def predict_proba(self,X):
+        proba = self.model.predict_proba(X)
+        return proba
+    
+    def get_predict(self,rows:str="all"):
+        """
+        Return predicted proba
+        """
+        if rows == "tagged":
+            return self.proba[self.df[self.col_label].notnull()]
+        if rows == "untagged":
+            return self.proba[self.df[self.col_label].isnull()]
+        
+        return self.proba
         
 def to_dfm(texts: Series,
            tfidf:bool=False,
